requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap-v5==1.0.7
django-grappelli==2.15.3
django-phonenumber-field==5.2.0
mysqlclient==2.1.0
phonenumberslite==8.12.33
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
